+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 
 - Fix issue where dropdown is hard to close on Firefox (#194, thanks to @dmarchuk).
 - Filter selected children from results (#201, thanks to @ne0guille).
+- Move submenu up if otherwise it would fall below the fold of the page.
 
 ## 3.0.4